+++ packages/web-console/src/scenes/Import/ImportCSVFiles/upload.tsx
@@ -42,7 +42,7 @@ const CopySQLLink = () => (
     target="_blank"
     rel="noopener noreferrer"
   >
-    here
+    COPY
   </a>
 )
 
@@ -104,6 +104,7 @@ export const Upload = ({ files, onFilesDropped, dialogOpen }: Props) => {
               }}
               prefixIcon={<Search2 size="18px" />}
               skin="secondary"
+              data-hook="import-browse-from-disk"
             >
               Browse from disk
             </Button>