+++ packages/web-console/src/utils/questdb/client.ts
@@ -214,7 +214,15 @@ export class Client {
       (response.ok && response.status === 403)
     ) {
       const fetchTime = (new Date().getTime() - start.getTime()) * 1e6
-      const data = (await response.json()) as RawResult
+      let data;
+      try {
+        data = (await response.json()) as RawResult
+      } catch (error) {
+        return Promise.reject({
+          error: `Invalid JSON response from the server: ${error}`,
+          type: Type.ERROR,
+        })
+      }
 
       eventBus.publish(EventType.MSG_CONNECTION_OK)
 
@@ -350,6 +358,7 @@ export class Client {
   async uploadCSVFile({
     file,
     name,
+    owner,
     settings,
     schema,
     partitionBy,
@@ -373,6 +382,7 @@ export class Client {
     const params = {
       fmt: "json",
       name,
+      owner,
       ...(partitionBy ? { partitionBy } : {}),
       ...(timestamp ? { timestamp } : {}),
       ...serializedSettings,